Formatting PR updates
* move fmt::Display impl for ResourceBlock to display mod
* remove now unnecessary serde derives for ShorthandRule
* remove `.to_polar` call in format string (implicit due to
  fmt::Display impl)
